# Improve performance with inline typecheck for string parsing